# List append optimizations for comprehension loops