--- a/common/types/traits/lister.go
+++ b/common/types/traits/lister.go
@@ -25,3 +25,8 @@ type Lister interface {
 	Iterable
 	Sizer
 }
+
+// MutableLister interface which emits an immutable result after an intermediate computation.
+type MutableLister interface {
+	ToImmutableList() Lister
+}