Introduce the list append optimizations used within C++ to ensure that comprehensions are able to treat internal
CEL variables as mutable for the purposes of list construction. Closes #488 

Before the optimization:
```
BenchmarkContextEval-8   	     250	   4680999 ns/op	   43948 B/op	    1323 allocs/op
```

After the optimization:
```
BenchmarkContextEval-8   	   10233	    116829 ns/op	   35516 B/op	    1038 allocs/op
```

The memory usage is not that different between the before and after, but the overall evaluation time (117us vs. 4.58ms) represents a 40x improvement in evaluation speed which will significantly help comprehension-heavy uses of CEL with even moderately sized inputs.